update / preparing to move everything to object literals (bc fml)
