+++ modules/botCommands.js
@@ -1,0 +1,6 @@
+
+module.exports = {
+    handleBotCommand: ()=>{
+        
+    }
+}
